Resolve circular import and fix some bugs
