+++ app/blog/preview/[...slug]/page.tsx
@@ -7,7 +7,7 @@ import Title from "../../../../components/Title";
 import {getPreviewPostData} from "../../../../lib/blog/loadPreview";
 import {fetchAllImageProps} from "../../../../lib/blog/imagePropsFetcher";
 
-import BlogMarkdown, {getPureCloudinaryPath} from "../../[...slug]/BlogMarkdown";
+import BlogMarkdown from "../../[...slug]/BlogMarkdown";
 
 import styles from '../../../../styles/blog/blog.module.scss';
 
@@ -16,6 +16,7 @@ import {formatReadTime} from "../../../../lib/blog/readTime";
 import {parseWithBudouX} from "../../../../lib/wordSplit";
 import PostAttributes from "../../../../components/blog/PostAttributes";
 import {createErrorArticle, ErrorablePost} from "../../../../lib/blog/loadPreview";
+import {getPureCloudinaryPath} from "../../../../lib/blog/getPureCloudinaryPath";
 
 type Props = {
   params: {
