+++ components/blog/TogglePageViewLink.tsx
@@ -1,11 +1,11 @@
 'use client';
 
-import {BlogPost} from "../../lib/blog/load";
 import {useRouter} from "next/router";
 import {faFileLines, faToiletPaper} from "@fortawesome/free-solid-svg-icons";
 import EntryButton from "./EntryButton";
 import {IconProp} from "@fortawesome/fontawesome-svg-core";
 import {usePathname} from "next/navigation";
+import BlogPost from "../../lib/blog/blogPost";
 
 export default function TogglePageViewLink ({post}: { post: BlogPost }) {
   const pathname = usePathname() ?? ''
