+++ components/blog/article-parts/HorizontalImages.tsx
@@ -1,8 +1,9 @@
 import {BlogImageData} from "../../../lib/blog/imagePropsFetcher";
 import BlogImage, {ImageCaption} from "../BlogImage";
 import React from "react";
-import {getPureCloudinaryPath, parseInlineMarkdown} from "../../../app/blog/[...slug]/BlogMarkdown";
+import {parseInlineMarkdown} from "../../../app/blog/[...slug]/BlogMarkdown";
 import {ArticleParts} from "../../../lib/blog/articleParts";
+import {getPureCloudinaryPath} from "../../../lib/blog/getPureCloudinaryPath";
 
 const HorizontalImages: ArticleParts = (content, entry, imageSize) => {
   const regex = new RegExp('^!\\[.*?\]\\(')
